api/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
sqlparse==0.4.4
celery==5.3.6
python3-nmap
pytest-playwright
django-cors-headers